% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_sim_core_1_4.R
\name{DAISIE_ONEcolonist_1_4}
\alias{DAISIE_ONEcolonist_1_4}
\title{Old version of DAISIE_ONEcolonist}
\usage{
DAISIE_ONEcolonist_1_4(time, island_spec, stt_table)
}
\arguments{
\item{time}{Total time of simulation}

\item{island_spec}{Matrix with runtime simulation data}

\item{stt_table}{Matrix with species-through-time information on immigrant,
anagenetic, cladogenetic and total number of species.}
}
\value{
DAISIE simulation output object
}
\description{
Old version of DAISIE_ONEcolonist
}
