% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ontogeny_sim_proc_b.R
\name{island_area}
\alias{island_area}
\title{Calculate island area at a given point in time}
\usage{
island_area(t, Apars, shape)
}
\arguments{
\item{t}{A numeric with current time of simulation at which area should
be calculated.}

\item{Apars}{A numeric vector with 4 elements:
\itemize{
\item{Apars[1]: The total time which the island is set to exist.}
\item{Apars[2]: The maximum (peak) area.}
\item{Apars[3]: The time at which the peak area will be achieved.}
\item{Apars[4]: The sharpness of the curve. Usually set to 1. Higher
  values result in steeper curves.}
}}

\item{shape}{A numeric either 0, for constant area, or 1, for a beta curve.}
}
\value{
Returns a numeric with area at time \code{t}, given the parametrs
}
\description{
Describe a quadratic relationship between time and area, or
  return a constant vale if constant area is assumed.
}
\examples{
area <- DAISIEtesting:::island_area(
  t = 4,
  Apars = c(10, 5000, 2, 1),
  shape = 1
)
}
