% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_sim_core_1_4.R
\name{DAISIE_ONEcolonist_1_4}
\alias{DAISIE_ONEcolonist_1_4}
\title{Old version of DAISIE_ONEcolonist}
\usage{
DAISIE_ONEcolonist_1_4(time, island_spec, stt_table)
}
\arguments{
\item{time}{stub}

\item{island_spec}{stub}

\item{stt_table}{stub}
}
\value{
descendants
}
\description{
Old version of DAISIE_ONEcolonist
}
